Seperate Files
Moved Constants, functions and more utilities into different files.
